requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
qrcode[pil]>=7.4
Pillow>=10.0
psycopg[binary]>=3.1
python-dotenv>=1.0
gunicorn>=22.0
